producer/requirements.txt
#numpy==1.18.1
kafka-python==2.0.1
